aclib::collections::linkedlist::singly

Struct LinkedList

pub struct LinkedList<T> { /* private fields */ }

Implementations

impl<T> LinkedList<T>

pub fn new() -> Self

O(1), return empty linked list

pub fn len(&self) -> usize

O(1), return length of the list

pub fn is_empty(&self) -> bool

O(1), return true if the list has no item, else return false

pub fn peek_head(&self) -> Option<Ref<'_, T>>

O(1), peek the top of the list

pub fn peek_tail(&self) -> Option<Ref<'_, T>>

O(1), peek the end of the list

pub fn push(&mut self, item: T)

O(1), add new item to the top of the list

pub fn pop(&mut self) -> Option<T>

O(1), return and remove head item

pub fn enqueue(&mut self, item: T)

O(1), add new item to the end of the list

pub fn dequeue(&mut self) -> Option<T>

O(1), return and remove head item

pub fn append(&mut self, other: &mut Self)

O(1), append the other linked list