aclib::graph::without_petgraph::structure

Struct AdjacencyList

pub struct AdjacencyList<Weight, Direction> { /* private fields */ }

weighted/unweighted and directed/undirected graph structure

Implementations

impl<D> AdjacencyList<Unweighted, D>

pub fn dijkstra<T: Copy + Ord + Unsigned>( &self, start: usize, goal: usize, ) -> (Option<T>, Vec<usize>)

O((n+m)log(n)), shortest path number of hops and its route, with dijkstra algorithm

impl<W: Copy + Add + Ord + Zero, D> AdjacencyList<Weighted<W>, D>

pub fn dijkstra(&self, start: usize, goal: usize) -> (Option<W>, Vec<usize>)

O((n+m)log(n)), shortest path distance and its route, with dijkstra algorithm

impl<W, D> AdjacencyList<W, D>

pub fn bfs(&self, start: usize) -> Vec<usize>

O(n), return breadth first search order

pub fn dfs(&self, start: usize) -> Vec<usize>

O(n), return depth first search order

impl AdjacencyList<Unweighted, Undirected>

pub fn new_unweighted_undirected( (n, _m): (usize, usize), e: &[(usize, usize)], ) -> Self

O(m), convert sequence of edges(unweighted, undirected) to adjacency list

impl AdjacencyList<Unweighted, Directed>

pub fn new_unweighted_directed( (n, _m): (usize, usize), e: &[(usize, usize)], ) -> Self

O(m), convert sequence of edges(unweighted, undirected) to adjacency list

impl<W: Copy> AdjacencyList<Weighted<W>, Undirected>

pub fn new_weighted_undirected( (n, _m): (usize, usize), e: &[(usize, usize, W)], ) -> Self

O(m), convert sequence of edges(weighted, undirected) to adjacency list

impl<W: Copy> AdjacencyList<Weighted<W>, Directed>

pub fn new_weighted_directed( (n, _m): (usize, usize), e: &[(usize, usize, W)], ) -> Self

O(m), convert sequence of edges(weighted, undirected) to adjacency list

impl<W: Copy, D> AdjacencyList<Weighted<W>, D>

pub fn weight(&self, u: usize, v: usize) -> W

O(1), get weight of edge

impl<W, D> AdjacencyList<W, D>

pub fn neighbors(&self, node: usize) -> &HashSet<usize>

O(1), get node’s neighbors set reference

pub fn nodes_len(&self) -> usize

O(1), number of nodes

pub fn add_node(&mut self) -> usize

O(1), add node and return its index

pub fn remove_node(&mut self, node: usize) -> Option<HashSet<usize>>

O(1), remove node and return its neighbors if possible
warning, remain node index is not changed